requirements.txt
# streamlit-drawable-canvas 0.9.3 uses image_to_url, removed in Streamlit 1.44+
streamlit>=1.32.0,<1.44.0
streamlit-drawable-canvas>=0.9.3
pdf2image>=1.17.0
pypdf>=4.0.0
img2table>=1.4.0
pytesseract>=0.3.13
pandas>=2.0.0
pillow>=10.0.0
